backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.34
psycopg2-binary==2.9.9
alembic==1.13.2
python-jose[cryptography]
slowapi
jinja2
email-validator
openpyxl
pandas
python-multipart
pytz
pyodbc
google-generativeai
apscheduler
pytest==8.3.2
httpx==0.27.0
flake8==7.1.1
black==24.8.0
bandit==1.7.9
